lint: split tests into multiple files
